+++ mayan/apps/authentication/dependencies.py
@@ -1,5 +1,5 @@
 from mayan.apps.dependencies.classes import PythonDependency
 
 PythonDependency(
-    module=__name__, name='django-auth-ldap', version_string='==5.2.0'
+    module=__name__, name='django-auth-ldap', version_string='==5.3.0'
 )