+++ mayan/apps/authentication_oidc/dependencies.py
@@ -1,5 +1,5 @@
 from mayan.apps.dependencies.classes import PythonDependency
 
 PythonDependency(
-    module=__name__, name='mozilla-django-oidc', version_string='==4.0.1'
+    module=__name__, name='mozilla-django-oidc', version_string='==5.0.2'
 )