+++ mayan/apps/fundraisers/dependencies.py
@@ -1,3 +1,3 @@
 from mayan.apps.dependencies.classes import PythonDependency
 
-PythonDependency(module=__name__, name='Markdown', version_string='==3.10')
+PythonDependency(module=__name__, name='Markdown', version_string='==3.10.2')