+++ mayan/apps/platforms/dependencies.py
@@ -4,10 +4,10 @@ PythonDependency(
     module=__name__, name='gevent', version_string='==25.9.1'
 )
 PythonDependency(
-    module=__name__, name='greenlet', version_string='==3.2.4'
+    module=__name__, name='greenlet', version_string='==3.3.1'
 )
 PythonDependency(
-    module=__name__, name='gunicorn', version_string='==23.0.0'
+    module=__name__, name='gunicorn', version_string='==25.1.0'
 )
 PythonDependency(
     module=__name__, name='whitenoise', version_string='==6.11.0'